refactor: :recycle: extract logic
